Keep compatible to Postgres

+++ src/main/resources/db/migration/V0_1_1__Create-story-table.sql
@@ -1,8 +1,8 @@
 CREATE TABLE Story (
     id SERIAL NOT NULL PRIMARY KEY,
-    version_major TINYINT, -- breaking changes
-    version_minor TINYINT, -- optional additions
-    version_patch TINYINT, -- cosmetic changes
+    version_major SMALLINT, -- breaking changes
+    version_minor SMALLINT, -- optional additions
+    version_patch SMALLINT, -- cosmetic changes
     about TEXT,
     language VARCHAR(2),
     -- tags separate table?
